Jump physics (#14)
* Check y-axis overlap in `collides?`

* Test gravity

Implement an initial test for gravity.

* Eliminate FloorSprite

And earlier implementation had a class called `FloorSprite`, which
was really just a wally sprite that happened to be at the bottom
of the screen. Now floors are represented by a wall sprite to reduce
code duplication.

* whitespace and rubocop

* rubo on test folder

* Start Jump Physics

Jumping now applies a single delta-y to the player sprite, which
then falls based on including the gravity module. Also fixed a bug
around y-coordinates going up towards the bottom of the screen.

* WIP commit for physics refactor

* Fix bug in nearest method

Previously, when a sprite checked the distance for its
nearest left and right neighbors, it would call a method
on nil and raise an error. Now it will instead return a
large integer.

* Fix sticking to top of things

Previously, the character sprite could be zero pixels off the top
of a wall, making the player unable to move left and right. This
change causes vertical collisions to occur at a pixel distance of
1 instead of zero.